[spatialPartitioning] Move KnnGraph in NeighborGraph folder
