+++ Ponca/SpatialPartitioning
@@ -15,4 +15,4 @@
 #include "src/SpatialPartitioning/KdTree/kdTreeTraits.h"
 #include "src/SpatialPartitioning/NeighborGraph/neighborGraph.h"
 #include "src/SpatialPartitioning/NeighborGraph/neighborGraphTraits.h"
-#include "src/SpatialPartitioning/KnnGraph/knnGraph.h"
+#include "src/SpatialPartitioning/NeighborGraph/knnGraph.h"
