+++ cmake/PoncaConfigureSpatialPartitioning.cmake
@@ -19,7 +19,7 @@ set(ponca_SpatialPartitioning_INCLUDE
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/NeighborGraph/Query/neighborGraphKNearestQuery.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/NeighborGraph/Query/neighborGraphRangeQuery.h"
     "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/NeighborGraph/Iterator/neighborGraphRangeIterator.h"
-    "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/KnnGraph/knnGraph.h"
+    "${PONCA_src_ROOT}/Ponca/src/SpatialPartitioning/NeighborGraph/knnGraph.h"
     )
 
 add_library(SpatialPartitioning INTERFACE)
