+++ tests/src/queries_nearest.cpp
@@ -16,7 +16,7 @@
 #include "../split_test_helper.h"
 
 #include <Ponca/src/SpatialPartitioning/KdTree/kdTree.h>
-#include <Ponca/src/SpatialPartitioning/KnnGraph/knnGraph.h>
+#include <Ponca/src/SpatialPartitioning/NeighborGraph/knnGraph.h>
 #include <Ponca/src/Common/pointTypes.h>
 
 #define PRINT_TIMING
